requirements.txt
Flask==2.1.1
Flask-RESTful==0.3.9
PyYAML==6.0
requests==2.27.1
openfoodfacts==0.1.3
nltk==3.7
inflect==5.5.2
